+++ internal/manager/run.go
@@ -16,6 +16,7 @@ import (
 	knativev1alpha1 "knative.dev/networking/pkg/apis/networking/v1alpha1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 	gatewayv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
@@ -59,7 +60,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 		return fmt.Errorf("get kubeconfig from file %q: %w", c.KubeconfigPath, err)
 	}
 	setupLog.Info("getting the kong admin api client configuration")
-	kongClients, err := c.getKongClients(ctx)
+	initialKongClients, err := c.getKongClients(ctx)
 	if err != nil {
 		return fmt.Errorf("unable to build kong api client(s): %w", err)
 	}
@@ -67,7 +68,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	// -------------------------------------------------------------------------
 
 	// Get Kong configuration root(s) to validate them and extract Kong's version.
-	kongRoots, err := kongconfig.GetRoots(ctx, setupLog, c.KongAdminInitializationRetries, c.KongAdminInitializationRetryDelay, kongClients)
+	kongRoots, err := kongconfig.GetRoots(ctx, setupLog, c.KongAdminInitializationRetries, c.KongAdminInitializationRetryDelay, initialKongClients)
 	if err != nil {
 		return fmt.Errorf("could not retrieve Kong admin root(s): %w", err)
 	}
@@ -79,13 +80,14 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	semV := semver.Version{Major: v.Major(), Minor: v.Minor(), Patch: v.Patch()}
 	versions.SetKongVersion(semV)
 
-	kongConfig := sendconfig.New(ctx, setupLog, kongClients, sendconfig.Config{
+	kongConfig := sendconfig.Config{
 		Version:            semV,
 		InMemory:           (dbMode == "off") || (dbMode == ""),
 		Concurrency:        c.Concurrency,
 		FilterTags:         c.FilterTags,
 		SkipCACertificates: c.SkipCACertificates,
-	})
+	}
+	kongConfig.Init(ctx, setupLog, initialKongClients)
 
 	setupLog.Info("configuring and building the controller manager")
 	controllerOpts, err := setupControllerOptions(setupLog, c, dbMode)
@@ -105,8 +107,22 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 
 	setupLog.Info("Initializing Dataplane Client")
 	eventRecorder := mgr.GetEventRecorderFor(KongClientEventRecorderComponentName)
-	dataplaneClient, err := dataplane.NewKongClient(
+
+	clientsManager, err := dataplane.NewAdminAPIClientsManager(
 		ctx,
+		deprecatedLogger,
+		initialKongClients,
+		adminapi.NewClientFactoryForWorkspace(c.KongWorkspace, c.KongAdminAPIConfig, c.KongAdminToken),
+	)
+	if err != nil {
+		return fmt.Errorf("failed to create AdminAPIClientsManager: %w", err)
+	}
+	if c.KongAdminSvc.Name != "" {
+		setupLog.Info("Running AdminAPIClientsManager notify loop")
+		clientsManager.RunNotifyLoop()
+	}
+
+	dataplaneClient, err := dataplane.NewKongClient(
 		deprecatedLogger,
 		time.Duration(c.ProxyTimeoutSeconds*float32(time.Second)),
 		c.IngressClassName,
@@ -116,7 +132,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 		kongConfig,
 		eventRecorder,
 		dbMode,
-		adminapi.NewClientFactoryForWorkspace(c.KongWorkspace, c.KongAdminAPIConfig, c.KongAdminToken),
+		clientsManager,
 	)
 	if err != nil {
 		return fmt.Errorf("failed to initialize kong data-plane client: %w", err)
@@ -150,7 +166,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 
 	setupLog.Info("Starting Enabled Controllers")
 	controllers, err := setupControllers(mgr, dataplaneClient,
-		dataplaneAddressFinder, udpDataplaneAddressFinder, kubernetesStatusQueue, c, featureGates, dataplaneClient)
+		dataplaneAddressFinder, udpDataplaneAddressFinder, kubernetesStatusQueue, c, featureGates, clientsManager)
 	if err != nil {
 		return fmt.Errorf("unable to setup controller as expected %w", err)
 	}
@@ -168,12 +184,7 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 	if err := mgr.AddHealthzCheck("health", healthz.Ping); err != nil {
 		return fmt.Errorf("unable to setup healthz: %w", err)
 	}
-	if err := mgr.AddReadyzCheck("check", func(_ *http.Request) error {
-		if !synchronizer.IsReady() {
-			return errors.New("synchronizer not yet configured")
-		}
-		return nil
-	}); err != nil {
+	if err := mgr.AddReadyzCheck("check", readyzHandler(mgr, synchronizer)); err != nil {
 		return fmt.Errorf("unable to setup readyz: %w", err)
 	}
 
@@ -194,8 +205,15 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 		// the argument checking the watch namespaces length enables or disables mesh detection. the mesh detect client
 		// attempts to use all namespaces and can't utilize a manager multi-namespaced cache, so if we need to limit
 		// namespace access we just disable mesh detection altogether.
-		if err := mgrutils.RunReport(ctx, kubeconfig, kongConfig, c.PublishService.String(), metadata.Release,
-			len(c.WatchNamespaces) == 0, featureGates); err != nil {
+		if err := mgrutils.RunReport(
+			ctx,
+			kubeconfig,
+			c.PublishService.String(),
+			metadata.Release,
+			len(c.WatchNamespaces) == 0,
+			featureGates,
+			clientsManager,
+		); err != nil {
 			setupLog.Error(err, "anonymous reporting failed")
 		}
 	} else {
@@ -204,6 +222,26 @@ func Run(ctx context.Context, c *Config, diagnostic util.ConfigDumpDiagnostic, d
 
 	setupLog.Info("Starting manager")
 	return mgr.Start(ctx)
+}
+
+type IsReady interface {
+	IsReady() bool
+}
+
+func readyzHandler(mgr manager.Manager, dataplaneSynchronizer IsReady) func(*http.Request) error {
+	return func(_ *http.Request) error {
+		select {
+		// If we're elected as leader then report readiness based on the readiness
+		// of dataplane synchronizer.
+		case <-mgr.Elected():
+			if !dataplaneSynchronizer.IsReady() {
+				return errors.New("synchronizer not yet configured")
+			}
+		// If we're not the leader then just report as ready.
+		default:
+		}
+		return nil
+	}
 }
 
 func getScheme() (*runtime.Scheme, error) {